Move redux into DynamicSourceFilters
